% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{add_county}
\alias{add_county}
\title{Add County Variable}
\usage{
add_county(df)
}
\arguments{
\item{df}{A data frame containing a countyIds column}
}
\value{
A data frame with a county factor column added
}
\description{
Adds a county name variable to the data frame based on county ID codes.
}
