% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_geometries}
\alias{add_geometries}
\title{Add traffic link geometries}
\usage{
add_geometries(traffic_data, id_col = "id")
}
\arguments{
\item{traffic_data}{Data frame with link id column}

\item{id_col}{Name of the id column (default "id")}
}
\value{
sf object with geometries joined
}
\description{
Add traffic link geometries
}
