% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{add_heavy_aadt}
\alias{add_heavy_aadt}
\title{Add heavy AADT}
\usage{
add_heavy_aadt(df)
}
\arguments{
\item{df}{A data frame containing total AADT and heavy ratio.}
}
\value{
A data frame with heavy vehicle AADT
}
\description{
Calculates heavy vehicle AADT from the heavy ratio.
}
