% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{add_logLastYear}
\alias{add_logLastYear}
\title{Add log-transformed length variable}
\usage{
add_logLastYear(df)
}
\arguments{
\item{df}{A data frame containing the column "lastYearAadt_aadt"}
}
\value{
A data frame with a "lastYearAadt_logAadt" and
"lastYearAadt_logHeavyAadt" column added.
}
\description{
Adds a log-transformed last year's AADT variable to the data frame.
}
