% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{add_logLength}
\alias{add_logLength}
\title{Add log-transformed length variable}
\usage{
add_logLength(df)
}
\arguments{
\item{df}{A data frame containing a length column}
}
\value{
A data frame with a logLength column added
}
\description{
Adds a log-transformed length variable to the data frame.
}
