% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{add_municipality}
\alias{add_municipality}
\title{Add Municipality Variable}
\usage{
add_municipality(df)
}
\arguments{
\item{df}{A data frame containing a municipalityIds column}
}
\value{
A data frame with a municipality factor column added
}
\description{
Adds a municipality name variable to the data frame based on municipality ID codes
using the municipality_names package data.
}
