% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{assign_traffic_volume_source}
\alias{assign_traffic_volume_source}
\title{Assign traffic volume source}
\usage{
assign_traffic_volume_source(df, current_year)
}
\arguments{
\item{df}{A data frame containing traffic volume data}

\item{current_year}{Integer specifying the current year}
}
\value{
A data frame with traffic_volume_source and traffic_volume_year columns added
}
\description{
Assigns a source label to each traffic volume measurement based on its characteristics.
}
