% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adjacency_matrix.R
\name{build_adjacency_matrix}
\alias{build_adjacency_matrix}
\title{Build adjacency matrix}
\usage{
build_adjacency_matrix(link_data, exclude_public_transport = FALSE)
}
\arguments{
\item{link_data}{Data frame with columns "startTrafficNodeId" and "endTrafficNodeId".}
}
\value{
Adjacency matrix. The number of rows and columns are both equal to the number of traffic links. An entry is 1 if the correspondng traffic links are connected, 0 otherwise.
}
\description{
Build adjacency matrix
}
