% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_incidence_matrix.R
\name{build_flow_constraints}
\alias{build_flow_constraints}
\title{Build flow constraint matrix}
\usage{
build_flow_constraints(link_data)
}
\arguments{
\item{link_data}{Data frame with columns "startTrafficNodeId" and "endTrafficNodeId".}
}
\value{
A matrix with rows equal to the number of nodes (with incoming and outgoing traffic), and columns equal to the number of traffic links. In a given row, the links corresponding to incoming traffic get value 1, and the links corresponding to outgoing traffic get value -1.
}
\description{
This is old and balances at traffic nodes, not flow nodes!!!
}
