% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurement_error.R
\name{build_measurement_matrix}
\alias{build_measurement_matrix}
\title{Build measurement matrix}
\usage{
build_measurement_matrix(data, colname_aadt = "aadt")
}
\arguments{
\item{data}{Data frame containing the row "aadt"}

\item{colname_aadt}{Column name of column containing AADT}
}
\value{
A matrix with each row corresponding to a traffic link with a measured traffic volume, and columns corresponding to all traffic links. For each row, the entry corresponding to the given traffic link has value 1.
}
\description{
Build measurement matrix
}
