% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bus_counts}
\alias{bus_counts}
\title{Bus stop counts}
\format{
A data frame with 149 rows and 10 variables.
\describe{
  \item{stopPointRef}{Identifier for the stop point}
  \item{no_of_buses}{The total number of buses that passed that bus stop.}
}
}
\usage{
data(bus_counts)
}
\description{
A data frame containing the number of buses at each bus stop throughout all of 2025. From EnTur. Raw, unprocessed data to test package functions on.
}
\keyword{data}
