% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buskerud_directed_traffic_links}
\alias{buskerud_directed_traffic_links}
\title{Directed traffic links for Buskerud}
\format{
A data frame with 1774 rows and 31 variables.
\describe{
  \item{id}{Unique identifier for the traffic link}
  \item{parentTrafficLinkId}{ID of the undirected traffic link}
  \item{isTrafficWithMetering}{Logical indicating if traffic direction is with metering or not}
  \item{functionalRoadClass}{Functional road classification of the traffic link}
  \item{functionClass}{Function class of the traffic link}
  \item{highestSpeedLimit}{Highest speed limit along the traffic link}
  \item{lowestSpeedLimit}{Lowest speed limit along the traffic link}
  \item{isNorwegianScenicRoute}{Logical indicating if this is a Norwegian scenic route}
  \item{isFerryRoute}{Logical indicating if this is a ferry route}
  \item{isRamp}{Logical indicating if this is a ramp}
  \item{isBlocked}{Logical indicating if the traffic link is blocked}
  \item{tollStationIds}{IDs of toll stations on this traffic link}
  \item{isInvalid}{Logical indicating if the traffic link data is invalid}
  \item{yearAppliesTo}{Year for which this traffic link data applies}
  \item{startTrafficNodeId}{ID of the node at the beginning of the traffic link}
  \item{endTrafficNodeId}{ID of the node at the end of the traffic link}
  \item{municipalityIds}{IDs of municipalities containing this traffic link}
  \item{countyIds}{IDs of counties containing this traffic link}
  \item{roadSystemReferences}{Road system references}
  \item{roadCategory}{Category of the road, e.g., E (European), R (national), F (county), K (municipal), P (private)}
  \item{roadLinkIds}{IDs of road links from NVDB}
  \item{roadNodeIds}{IDs of road nodes from NVDB}
  \item{roadPlacements}{Road placement information}
  \item{length}{Length of the traffic link in meters}
  \item{maxLanes}{Maximum number of lanes along the traffic link}
  \item{minLanes}{Minimum number of lanes along the traffic link}
  \item{hasOnlyPublicTransportLanes}{Logical indicating if the link has only public transport lanes}
  \item{associatedTrpIds}{IDs of associated traffic registration points}
  \item{lastYearAadt}{AADT information from the previous year (list column)}
  \item{bestDataSourceAadt}{AADT information from the best available data source (list column)}
  \item{trafficVolumes}{List containing traffic volume measurements and metadata}
}
}
\usage{
data(buskerud_directed_traffic_links)
}
\description{
Directed traffic links for Buskerud. Raw, unprocessed data to test package functions on.
}
\keyword{data}
