% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buskerud_nodes}
\alias{buskerud_nodes}
\title{Traffic nodes for Buskerud}
\format{
A sf data frame with 730 rows and 12 variables.
\describe{
  \item{id}{Unique identifier for the traffic node}
  \item{isRoundabout}{Logical indicating if the node is a roundabout}
  \item{numberOfIncomingLinks}{Number of directed traffic links entering this node}
  \item{numberOfOutgoingLinks}{Number of directed traffic links leaving this node}
  \item{numberOfUndirectedLinks}{Number of undirected traffic links at this node}
  \item{legalTurningMovements}{Legal turning movements allowed at this node}
  \item{connectedTrafficLinkIds}{List of undirected traffic link IDs connected to this node}
  \item{connectedTrafficLinkCandidateIds}{List of candidate traffic link IDs that connect to this node}
  \item{roadNodeIds}{IDs of road nodes from NVDB}
  \item{roadSystemReferences}{Road system references}
  \item{roadSystems}{Road systems connected to this node (indicates if there are connected road systems outside of the traffic links)}
  \item{geometry}{Spatial geometry of the node (point)}
}
}
\usage{
data(buskerud_nodes)
}
\description{
Traffic nodes for Buskerud. Raw, unprocessed data to test package functions on.
}
\keyword{data}
