% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_names}
\alias{county_names}
\title{County numbers and names}
\format{
A data frame with 357 rows and 2 variables.
\describe{
  \item{fylkesnavn}{County name}
  \item{fylkesnummer}{County number}
}
}
\usage{
data(county_names)
}
\description{
County numbers and names for all Norwegian counties
}
\keyword{data}
