% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_after_control.R
\name{estimate_standard_error}
\alias{estimate_standard_error}
\title{Estimate standard error for manual traffic estimates}
\usage{
estimate_standard_error(aadt)
}
\arguments{
\item{aadt}{Numeric vector of AADT values.}
}
\value{
Numeric vector of estimated standard errors.
}
\description{
Provides volume-dependent standard error estimates for manual AADT estimates
based on empirical uncertainty patterns.
}
\details{
Standard error bins:
\itemize{
  \item < 500: SE = 40
  \item 500-1000: SE = 100
  \item 1000-5000: SE = 200
  \item 5000-10000: SE = 400
  \item 10000-30000: SE = 1000
  \item 30000+: SE = 2000
}
}
\examples{
estimate_standard_error(c(200, 800, 3000, 15000, 40000))

}
