% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{extract_smallest_element}
\alias{extract_smallest_element}
\title{Extract smallest element}
\usage{
extract_smallest_element(x)
}
\arguments{
\item{x}{A vector or single value}
}
\value{
The minimum value if x is a vector, otherwise x itself
}
\description{
Extracts the smallest element from a vector, or returns the value if not a vector.
}
