% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_incidence_matrix.R
\name{find_connected_components}
\alias{find_connected_components}
\title{Find connected components in bipartite graph using union-find}
\usage{
find_connected_components(edges, all_incoming, all_outgoing)
}
\arguments{
\item{edges}{data frame with from/to columns representing graph edges}

\item{all_incoming}{character vector of incoming link IDs}

\item{all_outgoing}{character vector of outgoing link IDs}
}
\value{
list of character vectors, each representing a connected component
}
\description{
Find connected components in bipartite graph using union-find
}
