% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{flatten_df}
\alias{flatten_df}
\title{Flatten data frame}
\usage{
flatten_df(df)
}
\arguments{
\item{df}{A data frame with nested columns}
}
\value{
A flattened data frame
}
\description{
Flattens nested columns in the data frame, specifically bestDataSourceAadt and lastYearAadt.
}
