% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{get_best_traffic_volume}
\alias{get_best_traffic_volume}
\title{Get Best Traffic Volume}
\usage{
get_best_traffic_volume(df)
}
\arguments{
\item{df}{A data frame containing traffic volume data}
}
\value{
A processed data frame
}
\description{
Handles nested dataframes containing traffic volumes for data before 2024.
}
