% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_names}
\alias{municipality_names}
\title{Municipality numbers and names}
\format{
A data frame with 357 rows and 3 variables.
\describe{
  \item{kommunenavn}{Municipality name}
  \item{kommunenummer}{Municipality number}
  \item{fylke}{County}
}
}
\usage{
data(municipality_names)
}
\description{
Municipality numbers and names for all Norwegian municipalities.
}
\keyword{data}
