% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvdb_objects}
\alias{nvdb_objects}
\title{NVDB objects for Leaflet maps}
\usage{
nvdb_objects()
}
\value{
A list of nvdb_url, nvdb_attribution and nvdb_crs that are needed for using the NVDB map in Leaflet.
}
\description{
NVDB objects for Leaflet maps
}
\examples{
nvdb_objects()
}
