% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_traffic_links_simple_map}
\alias{plot_traffic_links_simple_map}
\title{Plot traffic links on an interactive map}
\usage{
plot_traffic_links_simple_map(
  link_data,
  color = "blue",
  color_by = NULL,
  palette = "RdYlGn",
  opacity = 0.8,
  weight = 3,
  popup_columns = NULL
)
}
\arguments{
\item{link_data}{A data frame containing at least an 'id' column with traffic
link IDs. Additional columns can be included for coloring and popups.}

\item{color}{Character string specifying the color for all links when `color_by`
is NULL. Default is "blue". Can be any valid CSS color name or hex code.}

\item{color_by}{Character string specifying the name of a column in `link_data`
to use for coloring the links. When NULL (default), all links use the same
color specified by `color`.}

\item{palette}{Character string specifying the color palette to use when
`color_by` is specified. Default is "RdYlGn". For continuous variables,
RdYlGn is automatically reversed so low values are green.}

\item{opacity}{Numeric value between 0 and 1 controlling the opacity of the
links. Default is 0.8.}

\item{weight}{Numeric value controlling the width of the link lines. Default is 3.}

\item{popup_columns}{Character vector specifying which columns from `link_data`
to include in the popup text. When NULL (default), only the link ID is shown.}
}
\value{
A leaflet map object that can be displayed or further customized.
}
\description{
Creates an interactive leaflet map displaying traffic links from the Norwegian
road database. Links can be displayed in a single color or colored by a variable.
}
\examples{
\dontrun{
# Simple map with all links in blue
my_links <- data.frame(id = c("12345", "67890"))
plot_traffic_links_simple_map(my_links)

# Color by a variable
my_links$road_type <- c("highway", "local")
plot_traffic_links_simple_map(my_links, color_by = "road_type")

# Custom popup with multiple columns
plot_traffic_links_simple_map(
  my_links,
  color_by = "road_type",
  popup_columns = c("id", "road_type")
)
}

}
