% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{preprocess_traffic_links}
\alias{preprocess_traffic_links}
\title{Preprocess data}
\usage{
preprocess_traffic_links(raw_directed_traffic_link_data, year = 2024)
}
\arguments{
\item{raw_directed_traffic_link_data}{Raw, unprocessed traffic count data}

\item{year}{The year of the data}
}
\value{
A processed data frame of directed traffic links.
}
\description{
Preprocess data
}
