% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{process_traffic_volume}
\alias{process_traffic_volume}
\title{Process traffic volume data}
\usage{
process_traffic_volume(df, year)
}
\arguments{
\item{df}{A data frame containing traffic volume data}

\item{year}{Integer specifying the year of the data}
}
\value{
A processed data frame with flattened traffic volume information
}
\description{
Processes traffic volume data based on the year, handling different data structures
for years before and after 2024.
}
