% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_incidence_matrix.R
\name{remove_incomplete_nodes}
\alias{remove_incomplete_nodes}
\title{Filter nodes based on balancing criteria}
\usage{
remove_incomplete_nodes(nodes, nodes_to_balance)
}
\arguments{
\item{nodes}{Data frame of traffic nodes}

\item{nodes_to_balance}{Either "all" or "complete nodes"}
}
\value{
Character vector of node IDs to include in flow balancing
}
\description{
Filter nodes based on balancing criteria
}
