% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{remove_list_columns}
\alias{remove_list_columns}
\title{Remove list columns}
\usage{
remove_list_columns(df)
}
\arguments{
\item{df}{A data frame potentially containing list columns}
}
\value{
A data frame with list columns removed
}
\description{
Removes all columns of type list from the data frame.
}
