% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{remove_redundant_columns}
\alias{remove_redundant_columns}
\title{Remove Redundant Columns}
\usage{
remove_redundant_columns(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with redundant columns removed
}
\description{
Removes redundant columns from the data frame.
}
