% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{round_and_check_aadt}
\alias{round_and_check_aadt}
\title{Round and Check AADT}
\usage{
round_and_check_aadt(df)
}
\arguments{
\item{df}{A data frame containing AADT values}
}
\value{
A data frame with rounded and checked AADT values
}
\description{
Rounds AADT values and ensures they are non-negative.
}
