% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{safely_extract_first_element}
\alias{safely_extract_first_element}
\title{Safely extract first element}
\usage{
safely_extract_first_element(x)
}
\arguments{
\item{x}{A vector, data frame, character, numeric, or NULL value}
}
\value{
The first element of x, or x itself if not a vector/data frame
}
\description{
Safely extracts the first element from a vector or data frame.
}
