% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_directed_traffic_links.R
\name{standardize_data_types}
\alias{standardize_data_types}
\title{Standardize Data Types}
\usage{
standardize_data_types(df)
}
\arguments{
\item{df}{A data frame with columns to be standardized}
}
\value{
A data frame with standardized column types
}
\description{
Converts columns to appropriate data types (numeric, integer, factor, logical).
}
