% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stops_on_traffic_links}
\alias{stops_on_traffic_links}
\title{Bus stops on traffic links}
\format{
A data frame with 149 rows and 10 variables.
\describe{
  \item{id}{Unique identifier for the directed traffic link}
  \item{roadSystemReferences}{Road system references}
  \item{onlyPtLastYear}{Logical indicating if this was also only public transport last year}
  \item{stopPointRef}{Identifier for the stop point}
  \item{stopCertainty}{Certainty level of the stop location, how certain is it that the number of buses at this stop represents the number of buses on the traffic link?}
  \item{stopOnTrafficLink}{Logical indicating if the stop is located on the traffic link}
  \item{stopAggregatesDirections}{Logical indicating if the stop aggregates both directions}
  \item{stopsServeDifferentBuses}{Logical indicating if the stops serve different bus routes}
  \item{comment}{Additional comments about the bus stop}
  \item{trikkestopp}{Logical indicating if this is a tram stop}
}
}
\usage{
data(stops_on_traffic_links)
}
\description{
A data frame containing all the traffic links that have only public transport lanes, with connected bus stops. Raw, unprocessed data to test package functions on.
}
\keyword{data}
