% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_incidence_matrix.R
\name{build_incidence_matrix}
\alias{build_incidence_matrix}
\title{Build incidence matrix from traffic links and constructed flow nodes}
\usage{
build_incidence_matrix(nodes, traffic_links, nodes_to_balance)
}
\arguments{
\item{nodes}{data frame containing legalTurningMovements column (which has JSON syntax)}

\item{traffic_links}{traffic link data set}
}
\value{
incidence matrix
}
\description{
Build incidence matrix from traffic links and constructed flow nodes
}
