% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_incidence_matrix.R
\name{determine_flow_type}
\alias{determine_flow_type}
\title{Determine the flow type based on incoming/outgoing link counts}
\usage{
determine_flow_type(incoming_links, outgoing_links)
}
\arguments{
\item{incoming_links}{character vector of incoming links}

\item{outgoing_links}{character vector of outgoing links}
}
\value{
character string describing flow type
}
\description{
Determine the flow type based on incoming/outgoing link counts
}
