% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_incidence_matrix.R
\name{find_connected_components}
\alias{find_connected_components}
\title{Find connected components using union-find algorithm}
\usage{
find_connected_components(edges, all_incoming, all_outgoing)
}
\arguments{
\item{edges}{Data frame with `from` and `to` columns representing graph edges}

\item{all_incoming}{Character vector of incoming link IDs}

\item{all_outgoing}{Character vector of outgoing link IDs}
}
\value{
List of character vectors, each containing node IDs in one component
}
\description{
Find connected components using union-find algorithm
}
